start to refactor helpers
