+++ contracts/test/hardhat/rollup/helpers/wallet.ts
@@ -1,0 +1,9 @@
+import { config, ethers } from "hardhat";
+import { HardhatNetworkHDAccountsConfig } from "hardhat/types";
+import { HDNodeWallet, Wallet } from "ethers";
+
+export const getWalletForIndex = (index: number) => {
+  const accounts = config.networks.hardhat.accounts as HardhatNetworkHDAccountsConfig;
+  const signer = HDNodeWallet.fromPhrase(accounts.mnemonic, "", `m/44'/60'/0'/0/${index}`);
+  return new Wallet(signer.privateKey, ethers.provider);
+};
